Trap system

+++ Assets/Prototyping/Scripts/Games Core/Mini games/Infinity Runner/PlayerRunnerManager.cs
@@ -6,6 +6,8 @@ namespace Prototyping.Games
 {
     public class PlayerRunnerManager : MonoBehaviour, IGamePortal
     {
+
+        [SerializeField] private LayerMask groundMask;
         [SerializeField] private Transform player;
         [SerializeField] private Transform startPosition;
 
@@ -15,6 +17,7 @@ namespace Prototyping.Games
         PlayerRunnerController controller;
 
         [SerializeField] private bool prototype;
+        [SerializeField] private bool runSystems;
 
         [SerializeField] private GameObject[] threeLaneBaseBlocks;
         [SerializeField] private float sideDashDistance;
@@ -23,6 +26,9 @@ namespace Prototyping.Games
         [SerializeField] private float movementSpeed;
         [SerializeField] private int totalBlockCount;
         [SerializeField] private Transform initialObject;
+
+        [SerializeField] private GameObject[] traps;
+        [SerializeField] private float trapGap;
 
         private bool gameStarted = false;
         private List<InfinityRunnerBlock> currentBlocks;
@@ -37,9 +43,12 @@ namespace Prototyping.Games
 
         void Update()
         {
-            MoveBlocks();
-            SpawnBlocks();
-            DeleteBlocks();
+            if (runSystems)
+            {
+                MoveBlocks();
+                SpawnBlocks();
+                DeleteBlocks();
+            }
         }
 
         void MoveBlocks()
@@ -59,17 +68,31 @@ namespace Prototyping.Games
             {
                 if (currentBlocks.Count < totalBlockCount)
                 {
-                    InfinityRunnerBlock lastGameObject = currentBlocks.Last();
-                    Vector3 lastGoEndPos = lastGameObject.End.position;
+                    InfinityRunnerBlock lastRunnerBlock = currentBlocks.Last();
+                    Vector3 lastGoEndPos = lastRunnerBlock.End.position;
                     GameObject gameObject = Instantiate(threeLaneBaseBlocks[Random.Range(0, threeLaneBaseBlocks.Length)], null, true);
                     InfinityRunnerBlock infinityRunnerBlock = gameObject.GetComponent<InfinityRunnerBlock>();
                     float offset = infinityRunnerBlock.End.position.z - infinityRunnerBlock.transform.position.z;
                     gameObject.transform.position = new Vector3(0f, yPosition, lastGoEndPos.z + offset);
-
-                    // infinityRunnerBlock.SetSidePosition(sideDashDistance);
-
                     currentBlocks.Add(infinityRunnerBlock);
+                    SpawnTrapsOnABlock(infinityRunnerBlock);
                 }
+            }
+        }
+
+        void SpawnTrapsOnABlock(InfinityRunnerBlock block)
+        {
+            float position = block.Start.position.z;
+            float offset = 0;
+            while (offset < block.Length - trapGap)
+            {
+                int trapPicker = Random.Range(0, traps.Length);
+                GameObject trap = Instantiate(traps[trapPicker], block.transform);
+                InfinityRunnerTrap trapScript = trap.GetComponent<InfinityRunnerTrap>();
+                float trapLength = trapScript.Length;
+                trap.transform.position = new Vector3(0f, block.Top.position.y + trapScript.YOffset, position + trapScript.Length / 2f);
+                position += trapLength / 2f + trapGap;
+                offset += trapLength / 2f + trapGap;
             }
         }
 
@@ -81,7 +104,7 @@ namespace Prototyping.Games
                 if (firstBlock != null)
                 {
                     InfinityRunnerBlock infinityRunnerBlock = firstBlock.GetComponent<InfinityRunnerBlock>();
-                    if (infinityRunnerBlock.End.position.z < player.position.z)
+                    if (infinityRunnerBlock.End.position.z < player.position.z - infinityRunnerBlock.Length)
                     {
                         currentBlocks.Remove(firstBlock);
                         Destroy(firstBlock.gameObject);
